pyproject.toml
[project]
name = "opokuml-geosight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy<2",
    "onnxruntime==1.18.0",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]
